requirements.txt
numpy
ipywidgets
webdavclient3
python-dotenv
scipy
sklearn
pandas
